common: extract MonoInfo class
